+++ src/main/java/be/ugent/piedcler/dodona/plugin/settings/SettingsPanel.form
@@ -11,7 +11,7 @@
     </clientProperties>
     <border type="none"/>
     <children>
-      <nested-form id="abfc3" form-file="be/ugent/piedcler/dodona/plugin/authentication/CredentialsPanel.form" binding="credentialsPanel" custom-create="true">
+      <nested-form id="abfc3" form-file="be/ugent/piedcler/dodona/plugin/ui/auth/CredentialsPanel.form" binding="credentialsPanel" custom-create="true">
         <constraints>
           <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
         </constraints>
