+++ src/main/java/be/ugent/piedcler/dodona/plugin/tasks/SelectExerciseTask.java
@@ -11,10 +11,10 @@ package be.ugent.piedcler.dodona.plugin.tasks;
 import be.ugent.piedcler.dodona.exceptions.DodonaException;
 import be.ugent.piedcler.dodona.plugin.Api;
 import be.ugent.piedcler.dodona.plugin.exceptions.UserAbortedException;
-import be.ugent.piedcler.dodona.plugin.ui.CourseSelectionDialog;
-import be.ugent.piedcler.dodona.plugin.ui.ExerciseSelectionDialog;
-import be.ugent.piedcler.dodona.plugin.ui.SelectionDialog;
-import be.ugent.piedcler.dodona.plugin.ui.SeriesSelectionDialog;
+import be.ugent.piedcler.dodona.plugin.ui.selection.CourseSelectionDialog;
+import be.ugent.piedcler.dodona.plugin.ui.selection.ExerciseSelectionDialog;
+import be.ugent.piedcler.dodona.plugin.ui.selection.SelectionDialog;
+import be.ugent.piedcler.dodona.plugin.ui.selection.SeriesSelectionDialog;
 import be.ugent.piedcler.dodona.resources.Course;
 import be.ugent.piedcler.dodona.resources.Exercise;
 import be.ugent.piedcler.dodona.resources.Series;
