+++ src/main/java/be/ugent/piedcler/dodona/plugin/ui/auth/CredentialsPanel.java
@@ -6,7 +6,7 @@
  *
  * https://github.com/thepieterdc/dodona-plugin-jetbrains
  */
-package be.ugent.piedcler.dodona.plugin.authentication;
+package be.ugent.piedcler.dodona.plugin.ui.auth;
 
 import be.ugent.piedcler.dodona.DodonaClient;
 import be.ugent.piedcler.dodona.exceptions.AuthenticationException;
