+++ src/main/java/be/ugent/piedcler/dodona/plugin/ui/auth/package-info.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2019. All rights reserved.
+ *
+ * @author Pieter De Clercq
+ * @author Tobiah Lissens
+ *
+ * https://github.com/thepieterdc/dodona-plugin-jetbrains
+ */
+
+/**
+ * UI components for authentication.
+ */
+
+@ParametersAreNonnullByDefault
+package be.ugent.piedcler.dodona.plugin.ui.auth;
+
+import javax.annotation.ParametersAreNonnullByDefault;
