+++ src/main/java/be/ugent/piedcler/dodona/plugin/ui/selection/CourseSelectionDialog.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="be.ugent.piedcler.dodona.plugin.ui.CourseSelectionDialog">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="be.ugent.piedcler.dodona.plugin.ui.selection.CourseSelectionDialog">
   <grid id="cbd77" binding="contentPane" layout-manager="BorderLayout" hgap="0" vgap="0">
     <constraints>
       <xy x="48" y="54" width="436" height="297"/>
