+++ src/main/java/be/ugent/piedcler/dodona/plugin/ui/selection/CourseSelectionDialog.java
@@ -6,7 +6,7 @@
  *
  * https://github.com/thepieterdc/dodona-plugin-jetbrains
  */
-package be.ugent.piedcler.dodona.plugin.ui;
+package be.ugent.piedcler.dodona.plugin.ui.selection;
 
 import be.ugent.piedcler.dodona.plugin.ui.listeners.SelectedItemListener;
 import be.ugent.piedcler.dodona.plugin.util.observable.ObservableValue;
