+++ src/main/java/be/ugent/piedcler/dodona/plugin/ui/selection/SeriesSelectionDialog.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="be.ugent.piedcler.dodona.plugin.ui.SeriesSelectionDialog">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="be.ugent.piedcler.dodona.plugin.ui.selection.SeriesSelectionDialog">
   <grid id="cbd77" binding="contentPane" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="10" left="10" bottom="10" right="10"/>
     <constraints>
