+++ src/main/java/be/ugent/piedcler/dodona/plugin/ui/selection/package-info.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2019. All rights reserved.
+ *
+ * @author Pieter De Clercq
+ * @author Tobiah Lissens
+ *
+ * https://github.com/thepieterdc/dodona-plugin-jetbrains
+ */
+
+/**
+ * UI components for selection.
+ */
+
+@ParametersAreNonnullByDefault
+package be.ugent.piedcler.dodona.plugin.ui.selection;
+
+import javax.annotation.ParametersAreNonnullByDefault;
